240918 오후11:30 커밋
- OAuth2, FormLogin시 액세스 토큰 및 리프레시 토큰 발급 완료
- 초기에 HTTP ONLY 쿠키에 담아서 전송. 이후 쿠키에 있는 액세스 토큰을 응답 헤더로 재전송 로직 완료
- 액세스 토큰 유효성 검사 로직 완료
- URL에 따라 JWT Filter 검증 유무 분리
- 로그아웃 시 토큰이 만료되어도 로그아웃 처리되게 처리
- TokenUtil을 통해 액세스 토큰을 통해 UserId 발급 로직 완료

+++ blog-springboot-jpa/src/main/java/com/yhs/blog/springboot/jpa/controller/PostApiController.java
@@ -1,5 +1,6 @@
 package com.yhs.blog.springboot.jpa.controller;
 
+import com.yhs.blog.springboot.jpa.config.jwt.TokenProvider;
 import com.yhs.blog.springboot.jpa.dto.PostRequest;
 import com.yhs.blog.springboot.jpa.dto.PostResponse;
 import com.yhs.blog.springboot.jpa.dto.PostUpdateRequest;
@@ -7,15 +8,22 @@ import com.yhs.blog.springboot.jpa.dto.PostUpdateResponse;
 import com.yhs.blog.springboot.jpa.entity.Post;
 import com.yhs.blog.springboot.jpa.service.PostService;
 import com.yhs.blog.springboot.jpa.util.PostMapper;
+import com.yhs.blog.springboot.jpa.util.TokenUtil;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 
 import java.security.Principal;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 @RequiredArgsConstructor
 @RestController
@@ -24,39 +32,71 @@ import java.util.List;
 public class PostApiController {
 
     private final PostService postService;
+    private final TokenProvider tokenProvider;
 
     @PostMapping
-    public ResponseEntity<PostResponse> createPost(@Valid @RequestBody PostRequest postRequest, Principal principal) {
+    public ResponseEntity<Object> createPost(@Valid @RequestBody PostRequest postRequest,
+                                             HttpServletRequest request) {
 
-        PostResponse responseDTO = postService.createPost(postRequest, principal);
-
+        // TokenAuthenticationFilter를 security에 등록해두었기 때문에, TokenAuthenticationFilter의
+        // SecurityContextHolder.getContext().setAuthentication(authentication)로 저장해둔 인증 정보를 가져옴
+        // 근데 TokenAuthenticationFilter에서 유효성 검사 통과하지 못하면 return 으로 모두 처리해두었기 때문에 불필요한 것 같아 주석 처리
+//        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+//        if (authentication == null || !authentication.isAuthenticated()) {
+//            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Unauthorized User");
+//        }
+        PostResponse responseDTO = postService.createPost(postRequest, request);
         return new ResponseEntity<>(responseDTO, HttpStatus.CREATED);
     }
 
     @GetMapping
     public ResponseEntity<List<PostResponse>> findAllPosts() {
-        List<PostResponse> postResponses = postService.getList().stream().map(post -> new PostResponse(post)).toList();
-
+        List<PostResponse> postResponses = postService.getPostList();
         return ResponseEntity.ok().body(postResponses);
 //        return new ResponseEntity<>(postResponses, HttpStatus.OK);
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<PostResponse> findPostById(@PathVariable("id") Long id) {
+    public ResponseEntity<PostResponse> findPostById(@PathVariable("id") Long postId) {
 
-        Post post = postService.getPost(id);
-        return ResponseEntity.ok().body(new PostResponse(post));
+        PostResponse postResponse = postService.getPost(postId);
+        return ResponseEntity.ok().body(postResponse);
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteArticle(@PathVariable("id") Long id) {
+    public ResponseEntity<String> deleteArticle(@PathVariable("id") Long id,
+                                           HttpServletRequest request) {
+
+//        TokenAuthenticationFilter에서 유효성 검사 실패하면 return 으로 모두 처리해두었기 때문에 불필요
+//        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+//        if (authentication == null || !authentication.isAuthenticated()) {
+//            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Unauthorized User");
+//        }
+
+        Long userId = TokenUtil.extractUserIdFromRequestToken(request, tokenProvider);
+        PostResponse postResponseDTO = postService.getPost(id);
+
+        if (!postResponseDTO.getUserId().equals(userId)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("You are not authorized to delete this post.");
+        }
+
         postService.deletePost(id);
         return ResponseEntity.ok().build();
     }
 
     //여기부터 작성
     @PatchMapping("/{id}")
-    public ResponseEntity<PostUpdateResponse> updateArticle(@PathVariable Long id, @RequestBody PostUpdateRequest postUpdateRequest) {
+    public ResponseEntity<Object> updateArticle(@PathVariable Long id,
+                                           @RequestBody PostUpdateRequest postUpdateRequest,
+                                                HttpServletRequest request) {
+
+        Long userId = TokenUtil.extractUserIdFromRequestToken(request, tokenProvider);
+        PostResponse postResponseDTO = postService.getPost(id);
+
+        if (!postResponseDTO.getUserId().equals(userId)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("You are not authorized to update this post.");
+        }
+
         Post updatedPost = postService.updatePost(id, postUpdateRequest);
         PostUpdateResponse postUpdateResponse = new PostUpdateResponse(updatedPost);
         return ResponseEntity.ok().body(postUpdateResponse);
